move FiscalYearSlider to guiComponent folder
