+++ frontend/src/components/guiComponents/FiscalYearSlider.vue
